KAFKA-7772: Move LoggingResource into resources package
