+++ checkstyle/import-control.xml
@@ -353,7 +353,7 @@
         <allow pkg="org.glassfish.jersey" />
         <allow pkg="com.fasterxml.jackson" />
         <allow pkg="org.apache.http"/>
-        <subpackage name="admin">
+        <subpackage name="resources">
           <allow pkg="org.apache.log4j" />
         </subpackage>
       </subpackage>
